[install][fix] Ignore readlink failures
Ignore the exit status of `readlink` when checking if `nvim_config_path`
is already a link.
When `nvim_config_path` does not exist or is not a link, `readlink`
exits with a failure and aborts the whole script because of `set -e`.
Use ` || true` to let the check always exit successfully.